reorganize manager inheritance
Move all fetch logic to an abstract AlgorithmManager that Commands and Questions extend. With tests
