+++ src/Commands/CommandNotFoundException.php
@@ -1,7 +1,8 @@
 <?php
-namespace Michaels\Midas\Exceptions;
+namespace Michaels\Midas\Commands;
 
 use Exception;
+use Michaels\Midas\Algorithms\AlgorithmNotFoundException;
 
 class CommandNotFoundException extends AlgorithmNotFoundException {}
