+++ src/Commands/InvalidCommandException.php
@@ -1,0 +1,8 @@
+<?php
+namespace Michaels\Midas\Commands;
+
+use Michaels\Midas\Algorithms\InvalidAlgorithmException;
+
+class InvalidCommandException extends InvalidAlgorithmException {}
+
+
