--- src/Exceptions/InvalidCommandException.php
@@ -1,6 +1,0 @@
-<?php
-namespace Michaels\Midas\Exceptions;
-
-class InvalidCommandException extends InvalidAlgorithmException {}
-
-
