+++ tests/unit/ManagerTest.php
@@ -6,52 +6,15 @@ use stdClass;
 
 class ManagerTest extends \PHPUnit_Framework_TestCase
 {
-    /**
-     * Test that true does in fact equal true
-     */
-    public function testAddByClassName()
+    public function testAddSingle()
     {
         $manager = new Manager();
-        $manager->add('classTest', 'Some\Class');
-        $items = $manager->getAll();
+        $manager->add('alias', 'value');
 
-        $this->assertArrayHasKey('classTest', $items, 'Array Items does not have key `classTest`');
-        $this->assertEquals('Some\Class', $items['classTest']);
+        $this->assertArrayHasKey('alias', $manager->getAll(), 'Array Items does not have key `objectTest`');
+        $this->assertEquals('value', $manager->get('alias'), 'Failed to get a single item');
     }
 
-    /**
-     * Test that true does in fact equal true
-     */
-    public function testAddItemByClosure()
-    {
-        $manager = new Manager();
-        $manager->add('closureTest', function() {
-            return true;
-        });
-
-        $items = $manager->getAll();
-
-        $this->assertArrayHasKey('closureTest', $items, 'Array Items does not have key `closureTest`');
-        $this->assertEquals(true, $items['closureTest']());
-    }
-
-    /**
-     * Test that true does in fact equal true
-     */
-    public function testAddItemByObject()
-    {
-        $manager = new Manager();
-        $manager->add('objectTest', new StdClass());
-
-        $items = $manager->getAll();
-
-        $this->assertArrayHasKey('objectTest', $items, 'Array Items does not have key `objectTest`');
-        $this->assertInstanceOf('StdClass', $items['objectTest'], 'Added object is not an object');
-    }
-
-    /**
-     * Test that true does in fact equal true
-     */
     public function testAddMultiple()
     {
         $manager = new Manager();
